+++ src/main/webapp/WEB-INF/contractinfo/contractinfolist.jsp
@@ -79,7 +79,7 @@
             <td>${contractinfo.verNo }</td>
             <td>${contractinfo.remark }</td>
             <td><a class="btn btn-success" href="contractinfo/updatePage/${contractinfo.id}">编辑</a>
-                <a id="delete" class="btn btn-primary" href="contractinfo/delete/${contractinfo.id}">删除</a>
+                <a id="delete" class="btn btn-primary" href="contractinfo/delete/${contractinfo.id}" onclick="{if(confirm('确认删除?')){return true;}return false;}">删除</a>
             </td>
         </tr>
     </c:forEach>
