+++ src/main/webapp/WEB-INF/menu/menulist.jsp
@@ -67,7 +67,7 @@ String basePath = request.getScheme()+"://"+request.getServerName()+":"+request.
               
              <td>
                 <a href="menu/updatePage/${u.menuId }">编辑</a>  
-                <a  href="menu/delete/${u.menuId }">删除</a>               
+                <a  href="menu/delete/${u.menuId }" onclick="{if(confirm('确认删除?')){return true;}return false;}">删除</a>
             </td>       
         </tr>	
         </c:forEach>
